Switched the cards from UI controls to Sprites
I initially used UI controls because it seemed like mouse input handling is simpler that way, and that is true, but I came to realize that there aren't good tools for other kinds of collision/overlap checks for UI controls, which meant that I would have either have had to whip up my own logic for that, or rely on the mouse events only. Using only mouse events might have been a reasonable choice, but it feels a bit limiting in terms of design, and doesn't match how Windows Solitaire handles it either, and since that's my inspiration for this project anyway, I wanted to be able to handle card overlap detection in a better way.

It does turn out using Sprites did mean I had to redo the input code, and some of it got a bit hairier than I'd like. In particular, I'm now keeping a list of cards the mouse is hovering over, and I'm checking their Z indices to determine if the card we're clicking on is the top one. This works, but relies on the fact that the cards have different Z indices to begin with, which is why in main.gd I've added setting the Z index when spawning the cards The cards also get their Z indices set when dragged on top of other cards, so that their Z index is one higher than the highest card it's piled on top of. This logic doesn't have a whole lot to do with how the actual game will play, but it does make for a neat enough toy for messing around making piles of cards.