Fixed some issues with stacking and Z indexing
The Z index wasn't being set correctly when a card was dragged but not dropped on a valid other card. I also found I had completely forgotten to set the Z index for cards added to the draw pile. While debugging these, I added a little debug label on the cards that shows their Z index. I'll get rid of it later but I kept it for now since it's useful while testing.

Adjusted the "pile offset" so that the cards are offset enough so you can read the numbers of cards even when other cards are on top of them.

There is a bug that might be new or might just be something I hadn't noticed before, where clicking on the deck starts dragging the card, which isn't an intended feature. I'm not sure how that's happening, but that's probably a good thing to fix next.