Added support for dragging sequences of cards, replenishing the deck
Dragging sequences was not too bad to do since I'm already using the parent/child relationship on the cards when they're piled up, but I did need to make sure that you can only drag sequences when they're in piles, and not accidental sequences in the draw pile for example, and I needed to make sure that you can't drop sequences of cards anywhere but other piles.

After I added this, I realized the way I use z_index needed to change since the child cards would need to be drawn on top of the parent card even when dragging. To faciliatet this, I changed the cards to use relative Z indexing, but since I still care about the "absolute index" to be able to ignore clicks on cards that are covered by other cards, I added a function that calculates the absolute index. This kind of feels like a feature that I'd expect to be a part of the engine but I couldn't find it. Simple enough logic anyway.

Added replenishing the deck once it's emptied. There's no limit to how many times you can do this at least for now. I changed the way drawing cards from the deck is handled for this, so the deck pile itself takes the click inputs instead of the top card... this resulted in the bug returning though where you can click and drag from the deck and the card will follow, which isn't intentional. This is somewhat random though since I think it depends on whether the click on the deck is handled before the click on the card. I'll need to solve that later.

Made the window resolution taller to avoid issues with card piles getting too tall. Not a real solution but good enough for now.

With these changes, the game is actually completable, though can take quite a few tries. Could use a restart/reshuffle button. This does also highlight for me the fact that Klondike is not the most satisfying solitaire to play, and I'm considering pivoting to a different game at this point since I have the basics down.