Fixed my terminology for "stack" and "pile" to "foundation" and "tableau"
Checked commonly accepted solitaire terminology, and it looks like the term for the piles where you're trying to move the cards into is "foundation", and the term for the piles of cards where you move cards around and most of the gameplay happens is "tableau", so both to get rid of the ambiguity I had between generic "pile" and the main play area "pile", so now I continue to use the generic "pile" but the main play area is "tableau". Also "stack" is pretty much a synonym for "pile" and also a data structure, so "foundation" is definitely better there.