Edited the card visuals to use distinct colors for each suit
Made diamonds blue and clubs green to help differentiate them better, as I was finding that when cards were piled up it was impossible to differentiate between suits of the same color, since the top of the card doesn't show the suit except on the face cards.  I chose hues that should be equally far apart from each other on the color wheel.