Changed the input code from using mouse events to touch events
I found that there was an issue with using mouse events for input, and in particular the way I was using the mouse entered and mouse exited events on the cards to collect a list of hovered cards, since on a touch screen there would be no hovering, so a touch input would sometimes happen before the "mouse" entered the card, making the inputs work inconsistently. I could have probably just worked around this, as I still have, and continued to use mouse events, but since my main platform is mobile, I decided to switch to using touch events altogether.

This has resulted in some overhead that wasn't necessary before, as it seems like the input_event signal on CollisionObject2D only seems to work on simulated touch inputs on the PC. This seems strange, because it did work for simulated mouse inputs on mobile. I'm trying to figure out if there's a way I could use those signals, as I'm not quite happy with the way my input code looks now. I think I may want to write my own centralized input handler script either way though, since I don't think there's a built-in method I could use to find the topmost card, for which I'm now using a physics point intersect query, so there's probably no way around that overhead, but right now my code is actually doing that query for every card when there's an input event, and that's quite wasteful. It does work though, and the game is now playable on mobile.