Moved card input code into its own script and node
The card input is now handled centrally in a "Card Input" node with its own script. This should be more efficient and feels less kludgey in general, as the collision queries are only done in one place. The logic is mostly the same, but now the currently dragged card is held in a variable instead of modifying the self object.