Added a simple Restart button
Added a simple Restart button to the UI. For now didn't move anything else as there was kind of a free spot next to the foundations, though I feel like that's not quite the right place for it. Also I may want to change it to a menu button also to avoid accidental presses, but for now this'll do I think. The restart button gets rid of all the existing cards on the board, and then redoes the game setup, which I've now extracted into its own function when it used to be in _ready.