Added detection for clicking/tapping a card instead of dragging it
Now I've postponed a bit of the dragging logic to the InputEventScreenDrag event, and if a touch is released without dragging being initiated, it counts as a click, and that emits a signal, which is now being handled by main.gd and it prints the name of the card. While I was at it I added a get_suit_name function so I can print out suit names more legibly.