Added auto move feature
Now when you click on a card, we check all the exposed cards and empty piles for valid moves, and if a valid move is found, the card is moved there. To accommodate this, I moved some of the logic of moving a card from one pile of another from card_input.gd to card.gd, and also I added a check that you can't move a card onto its own descendant. This was previously not necessary since it was already impossible to move a card onto another card that was being dragged.

The priority for which move to pick when multiple legal moves are available is such: 1. move to the card that was last moved (so an existing sequence won't be broken when moving multiple cards from a single pile) 2. move to a foundation 3. move to a card of the same suit 4, any other legal move