Added TouchScreenButton node for the Restart button
I didn't realize the Button node does not work with touchscreens, and I had been testing on PC so I didn't notice. I've added a TouchScreenButton node on top of it. This is a bit hacky, as the Button node is still there, but the TouchScreenButton doesn't have any visuals by default. I'll probably add some proper visuals for the button later, but for now the Button node acts as placeholder visuals for it. I removed the signal connection from the Button so now it's only the TouchScreenButton that responds to input, and this works both on PC and mobile since I have the PC emulating touch screen inputs.