Added a small movement threshold before dragging starts
Now a card does not start moving until it has been dragged at least 25 pixels. This is so that small movements will still register as a tap instead of a drag. 25 pixels seemed alright, making it feel more responsive to taps especially on mobile while not making it feel like the cards are sticky.