Updated the prioritization of auto-moves based on playtesting
I found that the "repeat the same move" logic sometimes felt weird if I had stopped to think for a second, or if I was moving cards from a whole different pile. Since the main purpose of this feature was to avoid cards starting to go all over the place when rapidly moving a whole stack, I added a little timer where a move is "repeated" only if it has been less than a second since the last move, and only for auto-moves. Manually dragging a card resets the stored previous move.

Also added a prioritization where moving a card onto another card is preferred over moving it onto an empty tableau.