Added little tweening animations for when cards are moved
I implemented this into Card.move_to and I updated the game setup and card draw code to use that. Added a little extra slowdown for when multiple cards are moved in one go at the game setup and card draw, so they move in a slightly staggered way. I did realize though that this doesn't quite work as nicely as I wanted it to because the cards are getting parented to the other newly added cards, and that makes them move together, which is not ideal for the staggered effect. For now I'm just accepting that it's not as nice as I'd like, but at least the columns get added in a staggered manner.