Added an Undo feature
in the main script, a move history is kept. A move is mainly "card X moved from card/pile Y to card/pile Z", but there's also the special case of drawing new cards, which counts as a single move, and also I store the information of whether a card was flipped right side up, so that will also be reversed when the undo is performed.

Added a new class Card.Move which holds the information about a single move. These objects are created both in Card.move_to which now returns a Card.Move, and in Main.deck_clicked, which generates a Card.Move for the card draw action.

The actual undo is handled in _on_undo_button_pressed which is connected to a signal from a touch button. For regular card moves, this basically moves the card back to where it was and also flips a card face down if one was flipped during that move. For the draw move, I iterate over the top cards of all the tableaus in reverse order so they end up in the original order, and I move each of those cards back into the deck, face down.